Fix typo in ephemeral env variable
The start script was checking for `EMPHEMERAL` rather than `EPHEMERAL`
and thus always setting the default value of `--ephemeral`.

I've also changed it to default to non ephemeral unless the caller
specifically sets the `EPHEMERAL` environment variable which sounds
clearer (so that if one doesn't want the runner to be ephemeral they
just don't set the environment variable rather than having to instead
set it to empty).

Note that this will change the behavior of the runner for existing
callers, not sure how you want to handle this.